RPG Maker v4.7.1 - CRITICAL UPDATE -
Patches an issue causing RPG Maker to try to load CONFIG_MAIN out of order, causing the RPG to randomly not load properly on certain servers.
Also removed the option to rename the /readyup/rpg/config.cfg file as it was causing problems, though this was a hidden/devops feature so no one should have been using it.

If you're on any version of v4.6.x please update to this build and any issues you're having should be resolved.